migration tahap 1
